pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thermal-palette"
version = "0.1.0"
description = "Thermal image palette alignment service"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26,<3",
    "Pillow>=10.0",
    "pandas>=2.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "typer>=0.12",
    "python-dateutil>=2.9",
]

[project.scripts]
thermal-palette = "thermal_palette.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.9",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
